fix: revert back to modular testing structure
